+++ tools/iana.py
@@ -36,10 +36,10 @@ import lxml.etree
 from httpolice.citation import RFC, Citation
 import httpolice.known
 from httpolice.structure import (AltSvcParam, AuthScheme, CacheDirective,
-                                 ContentCoding, FieldName, MediaType, Method,
-                                 Preference, RangeUnit, RelationType,
-                                 StatusCode, TransferCoding, UpgradeToken,
-                                 WarnCode)
+                                 ContentCoding, FieldName, ForwardedParam,
+                                 MediaType, Method, Preference, RangeUnit,
+                                 RelationType, StatusCode, TransferCoding,
+                                 UpgradeToken, WarnCode)
 from httpolice.util.text import normalize_whitespace
 
 
@@ -154,6 +154,7 @@ class ParametersRegistry(Registry):     # pylint: disable=abstract-method
         tree = self._get_xml('http-parameters/http-parameters.xml')
         return [
             (ContentCoding, list(self._content_codings(tree))),
+            (ForwardedParam, list(self._forwarded_parameters(tree))),
             (Preference, list(self._preferences(tree))),
             (RangeUnit, list(self._range_units(tree))),
             (TransferCoding, list(self._transfer_codings(tree))),
@@ -167,6 +168,18 @@ class ParametersRegistry(Registry):     # pylint: disable=abstract-method
                 '_': ContentCoding(
                     record.find('iana:name', self.xmlns).text),
                 '_citations': list(self.extract_citations(record)),
+            }
+
+    def _forwarded_parameters(self, tree):
+        records = tree.findall(
+            '//iana:registry[@id="forwarded"]/iana:record', self.xmlns)
+        for record in records:
+            yield {
+                '_': ForwardedParam(
+                    record.find('iana:name', self.xmlns).text),
+                '_citations': list(self.extract_citations(record)),
+                'description':
+                    record.find('iana:description', self.xmlns).text,
             }
 
     def _preferences(self, tree):